require lower-switch in preISel passes.
This is a workaround which is likely to be superseded by work in
progress on a better divergence analysis that can handle irreducible
blocks.

The default target of the switch instruction may sometimes be an
"unreachable" block, when it is guaranteed that one of the cases is
always taken. The dominator tree incorrectly computes that that such a
switch instruction does not have an immediate post dominator. This
confuses divergence analysis, which is unable to propagate sync
dependence to the targets of the switch instruction.

As a workaround, the AMDGPU target now includes lower-switch in the
list of preISel passes. The LowerSwitchPass handles the unreachable
default target correctly, allowing the dominator tree to find the
immediate post dominator.